requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
Django==4.2
django-filter==23.1
djangorestframework==3.14.0
dynaconf==3.1.12
hvac==1.1.0
idna==3.4
importlib-metadata==6.6.0
knox==0.1.14
loguru==0.7.0
Markdown==3.4.3
Pillow==9.5.0
pyhcl==0.4.4
pytz==2023.3
requests==2.29.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
validators==0.20.0
win32-setctime==1.1.0
zipp==3.15.0
